+++ app/src/main/java/woowacourse/shopping/data/remote/HeaderInterceptor.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.cart
+package woowacourse.shopping.data.remote
 
 import android.util.Base64
 import okhttp3.Interceptor
@@ -6,15 +6,18 @@ import okhttp3.Response
 
 class HeaderInterceptor(
     private val id: String,
-    private val password: String
+    private val password: String,
 ) : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         val credentials = "$id:$password"
         val basic = "Basic " + Base64.encodeToString(credentials.toByteArray(), Base64.NO_WRAP)
 
-        val newRequest = chain.request().newBuilder()
-            .header("Authorization", basic)
-            .build()
+        val newRequest =
+            chain
+                .request()
+                .newBuilder()
+                .header("Authorization", basic)
+                .build()
 
         return chain.proceed(newRequest)
     }
